Seperate authentication routes into a routing file
